respaldo/requirements.txt
cs50
Flask
Flask-Session
requests
tempfile
werkzeug.security
werkzeug.utils
datetime
os
flask-cors